[Service] Now saving Css and installing theme in saveTheme method of ThemeManager
The Css is now saved if the ‘css’ key is provided with $themeData in
the saveTheme method of ThemeManager.

The css key should be provided if `Cluckles` css export option is
enabled (See https://github.com/ilikeprograms/Cluckles) for that.

This is done to avoid server side compilation of the Css which can be
problematic on shared hosting.